Change remove partition to incorporate new check

+++ src/main/kotlin/dev/mizarc/bellclaims/application/actions/claim/partition/RemovePartition.kt
@@ -2,6 +2,7 @@ package dev.mizarc.bellclaims.application.actions.claim.partition
 
 import dev.mizarc.bellclaims.application.errors.DatabaseOperationException
 import dev.mizarc.bellclaims.application.persistence.PartitionRepository
+import dev.mizarc.bellclaims.application.results.claim.partition.CanRemovePartitionResult
 import dev.mizarc.bellclaims.application.results.claim.partition.RemovePartitionResult
 import java.util.UUID
 
@@ -10,7 +11,8 @@ import java.util.UUID
  *
  * @property partitionRepository Repository for managing partitions.
  */
-class RemovePartition(private val partitionRepository: PartitionRepository) {
+class RemovePartition(private val partitionRepository: PartitionRepository,
+                      private val canRemovePartition: CanRemovePartition) {
 
     /**
      * Removes the specified partition using its given [partitionId].
@@ -19,13 +21,20 @@ class RemovePartition(private val partitionRepository: PartitionRepository) {
      * @return An [RemovePartitionResult] indicating the outcome of the flag addition operation.
      */
     fun execute(partitionId: UUID): RemovePartitionResult {
-        // Add the flag to the claim
+        // Check if the removal will result in disconnection
+         when (canRemovePartition.execute(partitionId)) {
+             CanRemovePartitionResult.Success -> Unit
+             CanRemovePartitionResult.StorageError -> return RemovePartitionResult.StorageError
+             CanRemovePartitionResult.WillBeDisconnected -> return RemovePartitionResult.WillBeDisconnected
+        }
+
+        // Remove the partition from the claim
         try {
             return when (partitionRepository.remove(partitionId)) {
                 true -> RemovePartitionResult.Success
                 false -> RemovePartitionResult.DoesNotExist
             }
-        } catch (error: DatabaseOperationException) {
+        } catch (_: DatabaseOperationException) {
             println("Error has occurred trying to save to the database")
             return RemovePartitionResult.StorageError
         }
